fastapi/requirements.txt
-i https://pypi.org/simple
alembic==1.15.1; python_version >= '3.9'
annotated-types==0.7.0; python_version >= '3.8'
anyio==4.8.0; python_version >= '3.9'
click==8.1.8; python_version >= '3.7'
fastapi==0.115.11; python_version >= '3.8'
greenlet==3.1.1; python_version >= '3.7'
h11==0.14.0; python_version >= '3.7'
idna==3.10; python_version >= '3.6'
mako==1.3.9; python_version >= '3.8'
markupsafe==3.0.2; python_version >= '3.9'
psycopg2-binary==2.9.10; python_version >= '3.8'
pydantic==2.10.6; python_version >= '3.8'
pydantic-core==2.27.2; python_version >= '3.8'
python-dotenv==1.0.1; python_version >= '3.8'
sniffio==1.3.1; python_version >= '3.7'
sqlalchemy==2.0.38; python_version >= '3.7'
sqlmodel==0.0.23; python_version >= '3.7'
starlette==0.46.0; python_version >= '3.9'
typing-extensions==4.12.2; python_version >= '3.8'
uvicorn==0.34.0; python_version >= '3.9'
